able to merge audio processing
