+++ proto_redo/Assets/Scripts/ClapDetector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82579773c547f3c43b593672abd0cbe3
+guid: ac085da6351214147bd1ef7047d2b24d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
